monitor, fbdev: minor fixes

+++ display/fbdev.c
@@ -80,7 +80,7 @@ void fbdev_init(void)
 
     // Map the device to memory
     fbp = (char *)mmap(0, screensize, PROT_READ | PROT_WRITE, MAP_SHARED, fbfd, 0);
-    if((int)fbp == -1) {
+    if((intptr_t)fbp == -1) {
         perror("Error: failed to map framebuffer device to memory");
         return;
     }
